// src/app/domains/admin/modules/neuron/domain/sw42.enum.ts
export enum DocumentoPlantillaCaracteristicaEnum {
  TEXTO = 'T',
  FECHA = 'F',
  PROCESO = 'Z',
  NUMERO = 'N',
  BINARIO = 'I',
  PRODUCTO = 'J',
  ARCHIVO = 'A',
  CROQUIS = 'B',
  CONFIGURACION = 'G',
  DISPONIBILIDAD = 'U',
  PRODUCTO_LISTA = 'Q',
  SECCION = 'S',
  GPS = 'P',
  GPS_MAP = 'M',
  INFORMATIVE = 'V',
  VINCULO = 'C',
}

export enum StatesEnum {
  ACTIVE = 'A',
  INACTIVE = 'I',
  FINALIZADO = 'C',
}

export enum PropiedadEnum {
  ORGANIZACION = 'O',
  PROCESO = 'P',
  PLANTILLA = 'L',
  CAMPO = 'C',
  ESTADO = 'A',
  TRANSICION = 'T',
  REPORTE = 'E',
  ROL = 'R',
  API_SERVICE = 'W',
  SERVIDOR = 'S',
  CATALOG = 'G',
  ACCOUNT = 'K',
}

export enum FormatoCampoSimboloEnum {
  T = '\uD83D\uDCDD',
  F = '\uD83D\uDCC5',
  Z = '\uD83D\uDD04',
  N = '\uD83D\uDD22',
  I = '\u2699\uFE0F',
  J = '\uD83D\uDCE6',
  A = '\uD83D\uDCC1',
  B = '\uD83E\uDDE9',
  G = '\uD83D\uDEE0\uFE0F',
  U = '\uD83D\uDFE2',
  Q = '\uD83E\uDDEE',
  S = '\uD83D\uDCDA',
  P = '\uD83D\uDCCD',
  M = '\uD83D\uDDFA',
  V = '\uD83D\uDCA1',
  C = '\uD83D\uDD17',
}